Rematricula_modelo/requirements.txt
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0
sqlalchemy>=2.0
pyodbc>=5.0
python-dotenv>=1.0
joblib>=1.3
